Add Vitest and throttle test
